# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Barberia API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/root/create_owner": {
            "post": {
                "tags": [
                    "Root"
                ],
                "summary": "Crear propietario",
                "description": "Crea un nuevo propietario en la aplicación",
                "operationId": "CreateOwner",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "phone": {
                                        "type": "number",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Propietario registrado correctamente"
                    },
                    "400": {
                        "description": "Solicitud incorrecta"
                    },
                    "401": {
                        "description": "Debe verificar su correo electrónico para continuar."
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "422": {
                        "description": "Error de validación, verifique los campos"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Lista de usuarios",
                "description": "Muestra una lista de todos los usuarios",
                "operationId": "Usuarios",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios"
                    },
                    "401": {
                        "description": "El usuario no está verificado"
                    },
                    "500": {
                        "description": "Error en el servidor, Token inválido"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Los usuarios inician sesión en la aplicación y se genera un token para ellos, que les permite acceder a las rutas protegidas",
                "operationId": "Login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión correcto"
                    },
                    "400": {
                        "description": "Solictud incorrecta"
                    },
                    "401": {
                        "description": "El correo electrónico no está registrado"
                    },
                    "402": {
                        "description": "Contraseña incorrecta"
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "422": {
                        "description": "Error de validación, verifique los campos"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "description": "Crerra la sesión del usuario y elimina el token de autorización",
                "operationId": "Logout",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada"
                    },
                    "401": {
                        "description": "Debe verificar su correo electrónico para continuar."
                    },
                    "500": {
                        "description": "Error en el servidor, Token inválido"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        },
...
        "/api/auth/verify_email": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Verificar correo electrónico",
                "description": "Verifica el correo electrónico del usuario",
                "operationId": "verifyEmail",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "number",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "required": [
                                    "code"
                                ],
                                "properties": {
                                    "code": {
                                        "type": "number",
                                        "example": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Correo electrónico verificado correctamente"
                    },
                    "400": {
                        "description": "Código de verificación incorrecto"
                    },
                    "401": {
                        "description": "Debe verificar su correo electrónico para continuar."
                    },
                    "404": {
                        "description": "Recurso no encontrado"
                    },
                    "422": {
                        "description": "Error de validación, verifique los campos"
                    }
                },
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Ingresar el token de autorización",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer {token}",
                "scheme": "bearer"
            }
        }
    },
...
}